organize factory repo
